requirements.txt
uWSGI>=2.0.20,<2.1
Django>=4.1,<4.2
django-recaptcha>=3.0.0
firebrick>=2.0.0
pyyaml>=6.0
pillow>=9.2
requests>=2.28
tzdata>=2022.5
psycopg2-binary>=2.9.3